+++ qqrobot-spring-boot-starter/src/main/java/de/honoka/qqrobot/framework/impl/tester/model/TesterRobotMessage.java
@@ -9,8 +9,8 @@ import lombok.experimental.Accessors;
 import java.util.ArrayList;
 import java.util.List;
 
-@Data
 @Accessors(chain = true)
+@Data
 public class TesterRobotMessage {
 
     private List<Part> parts = new ArrayList<>();