+++ qqrobot-spring-boot-starter/src/main/java/de/honoka/qqrobot/starter/common/Exceptions.kt
@@ -1,8 +1,8 @@
 package de.honoka.qqrobot.starter.common
 
 class NoContactException(qq: Long? = null, group: Long? = null) : RuntimeException(buildString {
-    qq?.let { append("qq: $qq") }
-    group?.let { append(", group: $group") }
+    qq?.let { append("qq: $it") }
+    group?.let { append(", group: $it") }
 })
 
 class RobotMutedException(group: Long) : RuntimeException("group: $group")