+++ qqrobot-spring-boot-starter/src/main/java/de/honoka/qqrobot/starter/component/logger/entity/ExceptionRecord.java
@@ -6,8 +6,8 @@ import lombok.experimental.Accessors;
 import java.io.Serializable;
 import java.util.Date;
 
-@Data
 @Accessors(chain = true)
+@Data
 public class ExceptionRecord implements Serializable {
 
     private Integer id;